Shim changes
Added RapidsShuffleWriter and replaced 341db delta
dependency to 350db
