+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/RapidsShuffleInternalManagerBase.scala
@@ -38,7 +38,6 @@ import org.apache.spark.executor.ShuffleWriteMetrics
 import org.apache.spark.internal.{config, Logging}
 import org.apache.spark.io.CompressionCodec
 import org.apache.spark.network.buffer.ManagedBuffer
-import org.apache.spark.scheduler.MapStatus
 import org.apache.spark.serializer.SerializerManager
 import org.apache.spark.shuffle.{ShuffleWriter, _}
 import org.apache.spark.shuffle.api._
@@ -235,57 +234,6 @@ trait RapidsShuffleWriterShimHelper {
   def doCommitAllPartitions(writer: ShuffleMapOutputWriter, emptyChecksums: Boolean): Array[Long]
 }
 
-abstract class RapidsShuffleThreadedWriterBaseBase[K, V]()
-      extends ShuffleWriter[K, V]
-        with Logging {
-  protected var myMapStatus: Option[MapStatus] = None
-  protected val diskBlockObjectWriters = new mutable.HashMap[Int, (Int, DiskBlockObjectWriter)]()
-  /**
-   * Are we in the process of stopping? Because map tasks can call stop() with success = true
-   * and then call stop() with success = false if they get an exception, we want to make sure
-   * we don't try deleting files, etc twice.
-   */
-  private var stopping = false
-
-  def getMapStatus(
-    loc: BlockManagerId,
-    uncompressedSizes: Array[Long],
-    mapTaskId: Long): MapStatus = {
-    MapStatus(loc, uncompressedSizes, mapTaskId)
-  }
-
-  override def stop(success: Boolean): Option[MapStatus] = {
-    if (stopping) {
-      None
-    } else {
-      stopping = true
-      if (success) {
-        if (myMapStatus.isEmpty) {
-          // should not happen, but adding it just in case (this differs from Spark)
-          cleanupTempData()
-          throw new IllegalStateException("Cannot call stop(true) without having called write()");
-        }
-        myMapStatus
-      } else {
-        cleanupTempData()
-        None
-      }
-    }
-  }
-
-  private def cleanupTempData(): Unit = {
-    // The map task failed, so delete our output data.
-    try {
-      diskBlockObjectWriters.values.foreach { case (_, writer) =>
-        val file = writer.revertPartialWritesAndClose()
-        if (!file.delete()) logError(s"Error while deleting file ${file.getAbsolutePath()}")
-      }
-    } finally {
-      diskBlockObjectWriters.clear()
-    }
-  }
-}
-
 abstract class RapidsShuffleThreadedWriterBase[K, V](
     blockManager: BlockManager,
     handle: ShuffleHandleWithMetrics[K, V, V],
@@ -295,7 +243,7 @@ abstract class RapidsShuffleThreadedWriterBase[K, V](
     maxBytesInFlight: Long,
     shuffleExecutorComponents: ShuffleExecutorComponents,
     numWriterThreads: Int)
-      extends RapidsShuffleThreadedWriterBaseBase[K, V]
+      extends RapidsShuffleWriter[K, V]
         with RapidsShuffleWriterShimHelper {
   private val metrics = handle.metrics
   private val serializationTimeMetric =
@@ -1085,55 +1033,6 @@ abstract class RapidsShuffleThreadedReaderBase[K, C](
         new InterruptibleIterator[Product2[K, C]](context, resultIter)
     }
   }
-}
-
-abstract class RapidsCachingWriterBase[K, V](
-   blockManager: BlockManager,
-   handle: GpuShuffleHandle[K, V],
-   mapId: Long,
-   rapidsShuffleServer: Option[RapidsShuffleServer],
-   catalog: ShuffleBufferCatalog)
-  extends ShuffleWriter[K, V]
-    with Logging {
-  protected val numParts = handle.dependency.partitioner.numPartitions
-  protected val sizes = new Array[Long](numParts)
-
-  /**
-   * Used to remove shuffle buffers when the writing task detects an error, calling `stop(false)`
-   */
-  private def cleanStorage(): Unit = {
-    catalog.removeCachedHandles()
-  }
-
-  override def stop(success: Boolean): Option[MapStatus] = {
-    val nvtxRange = new NvtxRange("RapidsCachingWriter.close", NvtxColor.CYAN)
-    try {
-      if (!success) {
-        cleanStorage()
-        None
-      } else {
-        // upon seeing this port, the other side will try to connect to the port
-        // in order to establish an UCX endpoint (on demand), if the topology has "rapids" in it.
-        val shuffleServerId = if (rapidsShuffleServer.isDefined) {
-          val originalShuffleServerId = rapidsShuffleServer.get.originalShuffleServerId
-          val server = rapidsShuffleServer.get
-          BlockManagerId(
-            originalShuffleServerId.executorId,
-            originalShuffleServerId.host,
-            originalShuffleServerId.port,
-            Some(s"${RapidsShuffleTransport.BLOCK_MANAGER_ID_TOPO_PREFIX}=${server.getPort}"))
-        } else {
-          blockManager.shuffleServerId
-        }
-        logInfo(s"Done caching shuffle success=$success, server_id=$shuffleServerId, "
-          + s"map_id=$mapId, sizes=${sizes.mkString(",")}")
-        Some(MapStatus(shuffleServerId, sizes, mapId))
-      }
-    } finally {
-      nvtxRange.close()
-    }
-  }
-
 }
 
 class RapidsCachingWriter[K, V](
